add to lint ineffassign, intrange, modernize, tparallel, unconvert, usestdlibvars, usetesting (#5572)

+++ gossip/state/state.go
@@ -607,7 +607,7 @@ func (s *GossipStateProviderImpl) antiEntropy() {
 				continue
 			}
 
-			s.requestBlocksInRange(ourHeight, uint64(maxHeight)-1)
+			s.requestBlocksInRange(ourHeight, maxHeight-1)
 		}
 	}
 }
